Disable submit button while trix is uploading.
