+++ app/views/posts/_form.html.haml
@@ -1,9 +1,10 @@
 = form_with model: post, id: dom_id(post, "form"), 
-data: {controller: "action_text_whitelist"} do |f|
+data: { controller: "trix-form", 
+action: "trix-change->trix-form#checkTypeDisableSubmit" } do |f|
   = render 'notifications/error_messages', object: f.object
   = hidden_field_tag "repost[#{:post_id}]", @reposting.id if @reposting
   .field.mb-3
     = f.label :body, 'Compose a new post', class: "form-label"
     = f.rich_text_area :body, class: "form-control"
   .actions.mb-3
-    = f.submit class: 'btn btn-primary'
+    = f.submit class: 'btn btn-primary', data: { trix_form_target: "submit" }
